% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPcalibrate.R
\name{PPcalibrate}
\alias{PPcalibrate}
\title{Title}
\usage{
PPcalibrate(
  rc_determinations,
  rc_sigmas,
  calibration_curve,
  F14C_inputs = FALSE,
  n_iter = 1e+05,
  n_thin = 10,
  use_F14C_space = TRUE,
  show_progress = TRUE,
  sensible_initialisation = TRUE,
  calendar_ages = NA
)
}
\arguments{
\item{calendar_ages}{}
}
\description{
Title
}
